FEATURE ADDITION: CRTSCTS_Hardware_Flow_Control to Terminal Options.
All these patches are actually just to add this feature.

CRTSCTS_Hardware_Flow_Control => CRTSCTS

While this is NOT actually stated in the Ada POSIX Interface Specification, it's a modern option that is essential to serial operations for any modem or modem-like device.

I've only tested this on MacOS, but it works.  This option enables FULL hardware flow control, MacOS does have unidirectional form of this, I needed and used the normal bidirectional form.

Without this, buffer overrun is very common during binary transfer with a modem.  I added this for my own usage...but I'd prefer it be in the library natively.