Remove irrelevant dependency on libaio
When building florist on a Linux system with libaio installed, e.g.
on Debian with package libaio-dev, florist will be linked against
this library, even though it uses the aio_* routines provided by librt
and not the io_* routines provided by libaio.

This is similar to the recently fixed problem with irrelevant dependency
on libnsl.